Move auth to the core package
