+++ packages/core/test/auth.test.ts
@@ -1,0 +1,9 @@
+import auth from '../src/auth'
+
+describe('auth', () => {
+
+  test('.bearer()', () => {
+    expect(auth.bearer('==Qyahiadakkda')).toEqual({ scheme: 'bearer', credentials: '==Qyahiadakkda' } )
+  })
+  
+})
